Change handling of ... argument of curl_webhdfs
According to the documentation the ... argument in curl_webhdfs were passed
to getURL but the code actually passed them to curlOptions.
This rendered impossible the use of getURL options like .encoding or .mapUnicode.
I needed one of those options to enforce the correct encoding in a request that
was not handling the encoding correctly.